update_service view: handle `wait-for-index` query argument
with this argument set as `false`, the request should trigger an reindexing
request corresponding to the service update but not wait for its result.
though of course it has the same danger any async request has in that
overuse can cause a large number of requests to be sent to the target (in
this case the search api) in a very short amount of time. hopefully the
`update_service` view itself is slow enough to cause this to be throttled
to some extent.

this follows `user-role` (in this view) in being a Bit Weird because it's
accepting query string arguments in a POST request. but given the way we
use the body of a request implicitly as the contents of the new object,
supplying it there would be an even weirder option.