Allow overriding development server port and search api port
There are times when we might want to change the port that the search api
server listens on in development mode; this commit makes this easier by
adding an additional configuration variable `DM_SEARCH_API_PORT` that can
be set in the environment. We then construct the existing
`DM_SEARCH_API_URL` configuration variable from this.

The search api uses this same environment variable to determine what
port it listens on; see alphagov/digitalmarketplace-search-api#241. This
way we can change the entire system by changing one environment
variable.

This commit adds `DM_SEARCH_API_PORT` to the `.flaskenv` configuration
file; this works because `python-dotenv` does not override existing
environment variables.

We also use `DM_API_PORT` as the port for the api development server by
setting `FLASK_RUN_PORT`; see [the Flask cli docs][1] for more details
on how the development server is configured.

[1]: http://flask.palletsprojects.com/en/1.0.x/cli/